Remove test_climatology.py
